setup AddStudent form
configure AddStudent in /app/features/roster with validators/helpers. Form submits to rosterSlice in store, redirects to home on submit.
