+++ src/utils/formHelpers.js
@@ -1,0 +1,89 @@
+import * as yup from "yup";
+
+export const validationSchema = yup.object({
+  name: yup
+    .string("Enter your name")
+    .max(32, "Too many characters")
+    .matches(/^[aA-zZ\s]+$/, "Only alphabets are allowed for this field ")
+    .required("Name is required"),
+  city: yup
+    .string("Enter your city")
+    .max(32, "Too many characters")
+    .matches(/^[aA-zZ\s]+$/, "Only alphabets are allowed for this field ")
+    .required("City is required"),
+  zip: yup
+    .string("Enter your zip code")
+    .max(5, "Too many characters")
+    .matches(/^\d+$/, "Only numbers are allowed for this field ")
+    .required("Zip code is required"),
+});
+
+export const initialValues = {
+  name: "",
+  dateOfBirth: "",
+  city: "",
+  geoState: "",
+  zip: "",
+};
+
+export const statesArray = [
+  "Alabama",
+  "Alaska",
+  "American Samoa",
+  "Arizona",
+  "Arkansas",
+  "California",
+  "Colorado",
+  "Connecticut",
+  "Delaware",
+  "District of Columbia",
+  "Federated States of Micronesia",
+  "Florida",
+  "Georgia",
+  "Guam",
+  "Hawaii",
+  "Idaho",
+  "Illinois",
+  "Indiana",
+  "Iowa",
+  "Kansas",
+  "Kentucky",
+  "Louisiana",
+  "Maine",
+  "Marshall Islands",
+  "Maryland",
+  "Massachusetts",
+  "Michigan",
+  "Minnesota",
+  "Mississippi",
+  "Missouri",
+  "Montana",
+  "Nebraska",
+  "Nevada",
+  "New Hampshire",
+  "New Jersey",
+  "New Mexico",
+  "New York",
+  "North Carolina",
+  "North Dakota",
+  "Northern Mariana Islands",
+  "Ohio",
+  "Oklahoma",
+  "Oregon",
+  "Palau",
+  "Pennsylvania",
+  "Puerto Rico",
+  "Rhode Island",
+  "South Carolina",
+  "South Dakota",
+  "Tennessee",
+  "Texas",
+  "Utah",
+  "Vermont",
+  "Virgin Island",
+  "Virginia",
+  "Washington",
+  "West Virginia",
+  "Wisconsin",
+  "Wyoming",
+];
